(discovery, models): initial testing implementation of instance discovery. WARNING: THIS SERVICE CURRENTLY DOES NOT HAVE ANY AUTHENTICATION.
So far, there is no integration with the main api service, but this is an initial attempt to implement instance discovery. The current discovery service only offers a "CRUD"-like interface for instances directly inside Redis.

In its current state, it is very crude & unfinished, but I prefer to commit my work before I lose it.

The final implementation will likely rely on RedisJSON (ReJSON) & RediSearch.